requirements.txt
torch==2.4.0
transformers==4.45.0
accelerate==0.34.2
hydra-core==1.3.0
vllm==0.6.2
# flash-attn==2.4.3
wandb
datasets
nltk
rouge-score
git+https://github.com/jquesnelle/yarn.git@995db5b#egg=scaled-rope